+++ packages/branded/src/anemic.test.ts
@@ -7,12 +7,21 @@ import { __brand } from "./private-constants";
 import { AnemicOutput, BrandedType } from "./types";
 
 const Email = branded.primitive("Email", z.string().email());
-const [Address] = branded.shape("Address", z.object({ city: z.string(), street: z.string() }), {
-  methods: {},
-});
+const [Address] = branded.shape(
+  "Address",
+  z.object({
+    type: z.literal("Address").default("Address"),
+    city: z.string(),
+    street: z.string(),
+  }),
+  {
+    methods: {},
+  }
+);
 const [User] = branded.shape(
   "User",
   z.object({
+    type: z.literal("User").default("User"),
     email: branded.field(Email),
     verifiedAt: z.date().nullable(),
     address: branded.field(Address),